// src/navigations/types.tsx
export type RootStackParamList = {
  // Tabs: undefined;
  Splash: undefined;
  Home: undefined;
  Cart: undefined;
  Search: undefined;
  Profile: undefined;
  Like: undefined;
  ProductDetails: {
    image: string;
    title: string;
    price: number;
  };
};
export type DrawerParamList = {
  Tabs: undefined;
 Cart:undefined
};
export type BottomTabParamList = {
  Home: undefined;
  Products: undefined;
  Search: undefined;
  Profile: undefined;
  Like: undefined;
  // Like:undefined;
  Cart: undefined;
};